backend/control_center/requirements.txt
setuptools==75.8.2
absl-py==1.4.0
amqp==5.3.1
ansible==9.3.0
ansible-core==2.16.4
ansible-runner==2.3.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.1
blinker==1.8.1
cachetools==5.3.0
celery==5.4.0
certifi==2022.12.7
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==42.0.5
daphne==4.1.2
Deprecated==1.2.13
Django==5.0.1
django-celery==3.1.17
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-db-connection-pool[postgresql]==1.2.6
django-environ==0.12.0
django-filter==24.3
django-rest-knox==5.0.2
django-timezone-field==7.1
djangorestframework==3.14.0
docker==7.0.0
docopt==0.6.2
docutils==0.20.1
drf-spectacular==0.28.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
flatbuffers==23.1.21
gast==0.4.0
getmac==0.9.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka==1.3.5
keras==2.11.0
kombu==5.4.2
libclang==15.0.6.1
limits==3.3.1
lockfile==0.12.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgpack==1.0.8
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pexpect==4.9.0
pipreqs==0.4.11
prompt_toolkit==3.0.50
protobuf==3.19.6
psutil==5.9.8
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.36.2
requests==2.28.2
requests-oauthlib==1.3.1
resolvelib==1.0.1
rich==13.3.3
rpds-py==0.22.3
rsa==4.9
scapy==2.5.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn[standard]
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zope.interface==6.3